feat: solve day 16
as stated in the comment in part 1, I cheated for this one. I spent an
hour on my original attempt at a BFS which turned out to be a dead end
(haha maze joke) and I just gave up. I copied the solution from here:

https://github.com/JustSamuel/AOC/blob/9c7a7b1b55044120407df888480b5b75283b7beb/src/day16/index.ts

thank you @JustSamuel for the solution, I wouldn't have been able to get
this day without you

groetjes!