feat: solve day 17
Thanks to @hyperneutrino for help with part 2. The video (found here:
https://youtu.be/y-UPxMAh2N8) really helped me with figuring out how to
solve that one. Part 1 was really fun, part 2 too when you understand
what's going on.

I'm proud to say this is the first day this year I've really needed a
hyperneutrino video to figure out how to solve it. Next year I'll be
able to do it all on my own for sure definitely