feat: solve 2015 day 5
Had some trouble with part 2 for this one, where by splitting the string
and taking out the 2 characters we're searching for, sometimes it
becomes a valid string again. E.G. `qqaa` when the target is `qa`
becomes `q  a` -> `qa`, which happens to be the string we're searching
for. I've solved this in the end by just replacing the parts we're
searching for with `__`, so it would become `q__a`, which would not
trigger the search