feat: solve 2015 day 14
had some trouble with this one, mostly due to javascript
mutable/immutable weirdness and myself being not perfect with my
calculations. However, it was a funny one. In part 1 I optimized for a
massively long race time in part 2, so I precalculated the amount of
times a reindeer's cycle fits into the race time, and then simulating
the remainder. However, part 2 was just that you have to award a point
to the reindeer currently in the lead. So my (premature) optimization
was actually useless and I had to rewrite the whole thing to simulate
every tick anyway :D