feat: solve 2015 day 24
thanks @eddmann for the solution. This one was really tough. My initial
thought process was to generate all possible variants of the groups,
then pick the ones with the smallest first group and calculate the
Quantum Entanglement score of those, and return the smallest one of
that. However, I could not at all figure out how to get that done, and
other people's solutions quickly become so complicated it's hard to get
the specific functionality I needed.