feat: solve 2025 day 7
this was the first day where I was really thinking "Wow, I haven't done
this in a while". It was really hard for me to come up with the proper
solution for part 1, where my initial solution worked for the example
but not for the actual input. I've had to rewrite the whole thing maybe
3 times. Not very fun. Part 2 was easier luckily, so I was done with
that reasonably quickly